@@ -29,19 +29,20 @@
         }
     },
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "ext-bcmath": "*",
         "ext-exif": "*",
         "ext-intl": "*",
-        "laravel/framework": "^10.0",
-        "spatie/laravel-medialibrary": "^10.0.0",
+        "laravel/framework": "^10.0|^11.0",
+        "spatie/laravel-medialibrary": "^11.0.0",
         "spatie/laravel-activitylog": "^4.4",
         "laravel/scout": "^10.0",
-        "cartalyst/converter": "^8.0",
-        "kalnoy/nestedset": "^6.0",
+        "cartalyst/converter": "^8.0|^9.0",
+        "kalnoy/nestedset": "^v6.x-dev",
         "doctrine/dbal": "^3.6",
-        "spatie/laravel-blink": "^1.6",
         "lukascivil/treewalker": "0.9.1",
         "spatie/php-structure-discoverer": "^2.0"
+        "spatie/laravel-blink": "^1.7",
+        "lukascivil/treewalker": "0.9.1"
     }
 }
