@@ -36,7 +36,7 @@
     }
   },
   "require": {
-    "php": "^8.1",
+    "php": "^8.2",
     "lunarphp/lunar": "self.version"
   }
 }
