@@ -44,13 +44,7 @@ class AttributeData
         return $fieldType::getFilamentComponent($attribute)->label(
             $attribute->translate('name')
         )
-            ->formatStateUsing(fn ($state) => ($state ?: (new $attribute->type))->getValue())
-            ->dehydrateStateUsing(function ($state) use ($attribute) {
-                $field = new $attribute->type;
-                $field->setValue($state);
-
-                return $field;
-            })
+            ->formatStateUsing(fn ($state) => ($state ?: new $attribute->type))
             ->required($attribute->required)
             ->default($attribute->default_value);
     }
