@@ -77,6 +77,23 @@ class Attributes extends Forms\Components\Group
                     return $groupComponents;
                 },
             ]
-        )->configure()->key('attributeData');
+        )
+            ->configure()
+            ->key('attributeData')
+            ->mutateStateForValidationUsing(function ($state) {
+                if (! is_array($state)) {
+                    return $state;
+                }
+
+                foreach ($state as $key => $value) {
+                    if (! $value instanceof \Lunar\Base\Fieldtype) {
+                        continue;
+                    }
+
+                    $state[$key] = $value->getValue();
+                }
+
+                return $state;
+            });
     }
 }
