@@ -2,7 +2,7 @@
 
 namespace Lunar\Base;
 
-use Spatie\Image\Manipulations;
+use Spatie\Image\Enums\Fit;
 use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\MediaCollections\MediaCollection;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
@@ -13,7 +13,7 @@ class StandardMediaDefinitions implements MediaDefinitionsInterface
     {
         // Add a conversion for the admin panel to use
         $model->addMediaConversion('small')
-            ->fit(Manipulations::FIT_FILL, 300, 300)
+            ->fit(Fit::Fill, 300, 300)
             ->sharpen(10)
             ->keepOriginalImageFormat();
     }
@@ -60,7 +60,7 @@ class StandardMediaDefinitions implements MediaDefinitionsInterface
             foreach ($conversions as $key => $conversion) {
                 $model->addMediaConversion($key)
                     ->fit(
-                        Manipulations::FIT_FILL,
+                        Fit::Fill,
                         $conversion['width'],
                         $conversion['height']
                     )->keepOriginalImageFormat();
